% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{isbn_convert13to10}
\alias{isbn_convert13to10}
\alias{isbn_convert10to13}
\title{Convert ISBN-10 and ISBN-13}
\usage{
isbn_convert13to10(x)

isbn_convert10to13(x, .prefix = "978")
}
\arguments{
\item{x}{ISBN strings}

\item{.prefix}{For ISBN-13 prefix. Default is' '978'.}
}
\description{
Convert ISBN-10 and ISBN-13
}
\examples{
isbn_convert13to10("978-4-06-516404-4")
isbn_convert13to10(c("978-4-06-516404-4", "9784863542167"))

isbn_convert10to13(x = c("412345674X", "4022518286"))
}
